+++ src/columns_nansum_matrix_/index.d.ts
@@ -1,0 +1,4 @@
+export declare function columns_nansum_matrix_(
+	matrix:number[], columns_length:number
+):number[]
+export { columns_nansum_matrix_ as _matrix__nansum__columns }