+++ src/index.ts
@@ -1,7 +1,0 @@
-export * from './columns_nansum_matrix_.js'
-export * from './dot.js'
-export * from './dotMultiply.js'
-export * from './multiply_scalar_matrix.js'
-export * from './normalize.js'
-export * from './unit_matrix_.js'
-export * from './validate_length.js'