+++ src/mul_dot/index.d.ts
@@ -1,0 +1,1 @@
+export declare function mul_dot(...matrices:number[][]):number[]