+++ src/multiply_scalar_matrix/index.d.ts
@@ -1,0 +1,4 @@
+export declare function multiply_scalar_matrix(
+	scalar:number, matrix:number[]
+):number[]
+export { multiply_scalar_matrix as multiply__scalar__matrix }