+++ src/normalize/index.d.ts
@@ -1,0 +1,4 @@
+export declare function normalize(
+	matrix:number[], columns_length?:number
+):number[]
+export { normalize as normalize__row__major__matrix }