+++ src/unit_matrix_/index.d.ts
@@ -1,0 +1,1 @@
+export declare function unit_matrix_(length:number):number[]