+++ src/validate_length/index.d.ts
@@ -1,0 +1,2 @@
+export declare function validate_length(...arrays:any[][]):number
+export { validate_length as length__validate }