+++ src/validate_length/index.js
@@ -1,4 +1,8 @@
-export function validate_length(...arrays:any[][]) {
+/**
+ * @param arrays{any[][]}
+ * @returns {number}
+ */
+export function validate_length(...arrays) {
 	const length = arrays[0].length
 	for (let i = 1; i < arrays.length; i++) {
 		if (length != arrays[i].length) {